requirements.txt
awswrangler 
boto3 
python-decouple